Add store subfolders for 1st level state attributes.
Move actions inside to its related store subfolder.
