+++ src/store/blockchain/actions.ts
@@ -1,0 +1,176 @@
+import { BigNumber } from '0x.js';
+import { createAction } from 'typesafe-actions';
+
+import { TX_DEFAULTS, WETH_TOKEN_SYMBOL } from '../../common/constants';
+import { getContractWrappers } from '../../services/contract_wrappers';
+import { tokenToTokenBalance } from '../../services/tokens';
+import { getWeb3Wrapper, getWeb3WrapperOrThrow } from '../../services/web3_wrapper';
+import { getKnownTokens } from '../../util/known_tokens';
+import { BlockchainState, TokenBalance, Web3State } from '../../util/types';
+import { getOrderbookAndUserOrders, initializeRelayerData } from '../relayer/actions';
+import { getEthAccount, getTokenBalances, getWethBalance, getWethTokenBalance } from '../selectors';
+
+export const initializeBlockchainData = createAction('INITIALIZE_BLOCKCHAIN_DATA', resolve => {
+    return (blockchainData: BlockchainState) => resolve(blockchainData);
+});
+
+export const setEthAccount = createAction('SET_ETH_ACCOUNT', resolve => {
+    return (ethAccount: string) => resolve(ethAccount);
+});
+
+export const setWeb3State = createAction('SET_WEB3_STATE', resolve => {
+    return (web3State: Web3State) => resolve(web3State);
+});
+
+export const setTokenBalances = createAction('SET_TOKEN_BALANCES', resolve => {
+    return (tokenBalances: TokenBalance[]) => resolve(tokenBalances);
+});
+
+export const setEthBalance = createAction('SET_ETH_BALANCE', resolve => {
+    return (ethBalance: BigNumber) => resolve(ethBalance);
+});
+
+export const setWethBalance = createAction('SET_WETH_BALANCE', resolve => {
+    return (wethBalance: BigNumber) => resolve(wethBalance);
+});
+
+export const setWethTokenBalance = createAction('SET_WETH_TOKEN_BALANCE', resolve => {
+    return (wethTokenBalance: TokenBalance | null) => resolve(wethTokenBalance);
+});
+
+export const toggleTokenLock = ({ token, isUnlocked }: TokenBalance) => {
+    return async (dispatch: any, getState: any) => {
+        const state = getState();
+        const ethAccount = getEthAccount(state);
+
+        const contractWrappers = await getContractWrappers();
+
+        if (isUnlocked) {
+            await contractWrappers.erc20Token.setProxyAllowanceAsync(
+                token.address,
+                ethAccount,
+                new BigNumber('0'),
+                TX_DEFAULTS,
+            );
+        } else {
+            await contractWrappers.erc20Token.setUnlimitedProxyAllowanceAsync(token.address, ethAccount);
+        }
+
+        const isWeth = token.symbol === WETH_TOKEN_SYMBOL;
+        if (isWeth) {
+            const wethTokenBalance = getWethTokenBalance(state) as TokenBalance;
+            dispatch(
+                setWethTokenBalance({
+                    ...wethTokenBalance,
+                    isUnlocked: !isUnlocked,
+                }),
+            );
+        } else {
+            const tokenBalances = getTokenBalances(state);
+            const updatedTokenBalances = tokenBalances.map(tokenBalance => {
+                if (tokenBalance.token.address !== token.address) {
+                    return tokenBalance;
+                }
+
+                return {
+                    ...tokenBalance,
+                    isUnlocked: !isUnlocked,
+                };
+            });
+
+            dispatch(setTokenBalances(updatedTokenBalances));
+        }
+    };
+};
+
+export const updateWethBalance = (newWethBalance: BigNumber) => {
+    return async (dispatch: any, getState: any) => {
+        const state = getState();
+        const ethAccount = getEthAccount(state);
+        const wethTokenBalance = getWethTokenBalance(state);
+        const wethBalance = getWethBalance(state);
+
+        const web3Wrapper = await getWeb3WrapperOrThrow();
+        const networkId = await web3Wrapper.getNetworkIdAsync();
+        const wethAddress = getKnownTokens(networkId).getWethToken().address;
+
+        const contractWrappers = await getContractWrappers();
+
+        let tx: string;
+        if (wethBalance.lessThan(newWethBalance)) {
+            tx = await contractWrappers.etherToken.depositAsync(
+                wethAddress,
+                newWethBalance.sub(wethBalance),
+                ethAccount,
+            );
+        } else if (wethBalance.greaterThan(newWethBalance)) {
+            tx = await contractWrappers.etherToken.withdrawAsync(
+                wethAddress,
+                wethBalance.sub(newWethBalance),
+                ethAccount,
+                TX_DEFAULTS,
+            );
+        } else {
+            return;
+        }
+
+        await web3Wrapper.awaitTransactionSuccessAsync(tx);
+        const ethBalance = await web3Wrapper.getBalanceInWeiAsync(ethAccount);
+        dispatch(setEthBalance(ethBalance));
+
+        const newWethTokenBalance = wethTokenBalance
+            ? {
+                  ...wethTokenBalance,
+                  balance: newWethBalance,
+              }
+            : null;
+
+        dispatch(setWethTokenBalance(newWethTokenBalance));
+    };
+};
+
+export const initWallet = () => {
+    return async (dispatch: any) => {
+        dispatch(setWeb3State(Web3State.Loading));
+
+        const web3Wrapper = await getWeb3Wrapper();
+
+        if (web3Wrapper) {
+            const [ethAccount] = await web3Wrapper.getAvailableAddressesAsync();
+            const networkId = await web3Wrapper.getNetworkIdAsync();
+
+            const knownTokens = getKnownTokens(networkId);
+
+            const tokenBalances = await Promise.all(
+                knownTokens.getTokens().map(token => tokenToTokenBalance(token, ethAccount)),
+            );
+
+            const wethToken = knownTokens.getWethToken();
+            const wethTokenBalance = await tokenToTokenBalance(wethToken, ethAccount);
+
+            const ethBalance = await web3Wrapper.getBalanceInWeiAsync(ethAccount);
+
+            const selectedToken = knownTokens.getTokenBySymbol('ZRX');
+
+            dispatch(
+                initializeBlockchainData({
+                    web3State: Web3State.Done,
+                    ethAccount,
+                    ethBalance,
+                    wethTokenBalance,
+                    tokenBalances,
+                }),
+            );
+            dispatch(
+                initializeRelayerData({
+                    orders: [],
+                    userOrders: [],
+                    selectedToken,
+                }),
+            );
+            dispatch(getOrderbookAndUserOrders());
+        } else {
+            dispatch(setWeb3State(Web3State.Error));
+        }
+    };
+};
