+++ src/store/blockchain/reducers.ts
@@ -1,0 +1,42 @@
+import { BigNumber } from '0x.js';
+import { getType } from 'typesafe-actions';
+
+import { BlockchainState, Web3State } from '../../util/types';
+import * as actions from '../actions';
+import { RootAction } from '../common/reducers';
+
+const initialBlockchainState: BlockchainState = {
+    ethAccount: '',
+    web3State: Web3State.Loading,
+    tokenBalances: [],
+    ethBalance: new BigNumber(0),
+    wethTokenBalance: null,
+};
+
+export function blockchain(state: BlockchainState = initialBlockchainState, action: RootAction): BlockchainState {
+    switch (action.type) {
+        case getType(actions.setEthAccount):
+            return { ...state, ethAccount: action.payload };
+        case getType(actions.setWeb3State):
+            return { ...state, web3State: action.payload };
+        case getType(actions.setTokenBalances):
+            return { ...state, tokenBalances: action.payload };
+        case getType(actions.setWethTokenBalance):
+            return { ...state, wethTokenBalance: action.payload };
+        case getType(actions.setWethBalance):
+            return {
+                ...state,
+                wethTokenBalance: state.wethTokenBalance
+                    ? {
+                          ...state.wethTokenBalance,
+                          balance: action.payload,
+                      }
+                    : null,
+            };
+        case getType(actions.setEthBalance):
+            return { ...state, ethBalance: action.payload };
+        case getType(actions.initializeBlockchainData):
+            return action.payload;
+    }
+    return state;
+}
